Use sg_element as a POD class to add elements to a plot
The sg_element is just a shallow place-holder, a POD class, to indicate
the stroke, fill properties and the flags for plotting. It will be copied
as needed. The sg_element contain a naked pointer to an sg_object but
does not take its ownership. This latter belongs to the plot that accepts
the sg_object.

the graphics::path object is now an object that provides the sg_object
interface by simple inheritance and is not related to sg_element.