Use custom send event on xwindow to request a region update
Instead of calling update_region directly from another thread it does
send a custom event to the window's thread so that the drawing operation
is done by the window's thread.

Breaks the Windows implementation.

It does require XInitThread() because XSendEvent and XFlush is used without
mutexes from outside the window's thread.