Initial work on plot_legends using layouts
The idea is to represent legends as a list of pair of graphical elements
(the symbol) and text. Each element of the list will be rendered as a line
in the legend. The rendering will use the same approach of x and y axis
rendering by adjusting the plot layout to accommodate the text.

Depending on the legend placement to legend entries can be placed along a
column (if the legend is on the left) or in a row (if it is in the bottom).

In addition legend layout can be adjusted to display for examples in two
columns if there are too many entries.