Fix plot agent to work with and window surface
Now plot_agent exists in the private graphics namespace and
works directly with private implementation classes.

Instead of taking a plot to a Window object now plot_agent
takes a pointer to a window_surface object to call the
slot_refresh method. Every window has a window_surface as a
member variable so that plot_agent can obtain a point to a
window_surface by calling on the window the method
window_surface().

With this new implementation now automatic window update works
also for the FOX implementation when using a GraphicsWindow
inside a classic FOX application.