Configure install for C library and merge headers
Add a pkg-config file for the C interface library. There is currently a
problem in that to link an executable with gcc the option -lstdc++ is
needed but it is not included in the flags provided by pkg-config.

Now libcanvas.h will auto-detect if it is C++ or C and includes either
the C++ or C version of the public interface.