Implement separate auto limits for x and y axis
The logic with push/pop layer should be verified and maybe adjusted.

One inconvenient of the current implementation is that if the X axis
limits are fixed the automatic limits for the y axis will be based on
object extension along the y axis even for parts in the region of x axis
that are not plotted.

There is currently a bug with plot limits when AutoLimits is not
enabled.