Transform graphics::window into a pure interface
graphics::window_platform_native now provide an actual implementation
of a native window that complies to the graphics::window interface.

This should open the way to other non-native functions that can possibly
return a graphics::window which is not provided by window_platform_native.

Note that the elp::Window class of the public C++ interface only take a
pointer to graphics::window (named WindowImpl) so it can accept and use
any pointer of a class that provides the graphics::window interface.