Rename graphics::window to elp_window and simplify pimpl declarations
Now use directly elp_window as a WindowImpl for pimpl in Window's class.

The most important implication of this is that now WindowImpl cannot be
whatever the Window's implementation choose it to be but it needs to be
precisely a elp_window interface.

Because of this problem I was forced to "fix" FXElpWindow to declare
window_fox as its window implementation pimpl but this is actually more
logical.

Since it is now incompatible I have removed the FOX backend to provide
the native windows.