Factor base class window_fox_kernel from window_fox
The base class use a generic drawable and selector for "update region"
messages and does not provide a window_surface.

One of the difference from window_fox is that window_fox_kernel can be
created withoud a FOX application or drawable. The FXDrawable can be
attached later using the "bind_drawable" method.