Modify window_win32 and xwindow to provide a non-blocking start method
Assume now that the start method is non-blocking. Provide an helper
function start_window_new_thread to start a window creating a new thread
using the "start_blocking" method.

This is done because a FOX window, using window_fox_kernel, should
provide naturally a non-blocking start method because the window is
actually started in an already-existing FOX thread.