Add start method in window_fox_kernel class
Currently not used but should allow in principle to "start" a window
at the condition that a link is done with a FOX object capable of
creating a window. The function to use is "bind_window_environment" and
the mechanism is a FOX GUI signal sent to the target object with a
specific selector.