Re-implement FXElpWindow class to use FXElemPlotWindow
At this point the FXElpWindow class was redundant. Now it use the
FXElemPlotWindow to provide the FOX Window support and include directly
a Window instance to provide the library window's public C++ API
methods.

A special constructor is added in window_fox_kernel to create a window
that is not "started" in a separate thread but is directly linked to
the FXElemPlotWindow instance.