Use a FXGUISignal to create a FOX Window in a host environment
The FXGUISignal should be created in the GUI thread and not in the worker
thread. This was the GUI thread correctly awake when the signal is sent.

As a side effect the implementation is slightly simplified and more
efficient, only one signal for GUI thread is created and not one for each
created window.