Add a mechanism to retry update region requests
Seems to be overkill and not needed on Windows but it may be needed on
XWindow.

There is still a problem, when doing a slot_refresh() call, between the
call to update_request() and the call to commit_pending_draw() the
GUI thread can resize the window and the graphical updates will be lost.
It is a logical problem but there is no memory fault or race condition.