WIP: use SDL Renderer with Retina display support
It works on Linux but it needs more testing and some cleaning and should
be tested on macOS.

Use a SDL_Renderer and SDL_Texture to present the window_surface image
on the window.

The function SDL_GL_GetDrawableSize is used to get the real size of the
underlying buffer of the window. The size of the buffer can be greated,
by a integer factor, of the size of the window requested when creating
the window. Once the buffer size estabilished the window surface will
use the pixel units of the underlying buffer and never be aware of the
requested size of the window during its creation.